change test UI, create User, change Api, fixed bags
