+++ README.md
@@ -1,0 +1,13 @@
+# Тестирование Яндекс почты
+
+## Тестовые модули
+
+* [API module](.\moduleApi)
+
+* [UI module](.\moduleUi)
+
+* [Utils модуль](.\moduleUtils)
+
+### [Подготовка к запуску](./PREREQUISITES.md)
+
+### [Запуск тестов](./TESTORUN.md)
