+++ moduleApi/src/main/java/apiUtils/modelMessage/Field.java
@@ -1,0 +1,43 @@
+package apiUtils.modelMessage;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Field {
+    public String type;
+    public String email;
+    public String name;
+
+    @Override
+    public String toString() {
+        return "Field{" +
+                "type='" + type + '\'' +
+                ", email='" + email + '\'' +
+                ", name='" + name + '\'' +
+                '}';
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
