+++ moduleUtils/pom.xml
@@ -2,17 +2,16 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>yandexTest</artifactId>
+        <groupId>yulia.ipr</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>yulia.ipr</groupId>
-    <artifactId>yandexTest</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>moduleUtils</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <modules>
-        <module>moduleApi</module>
-        <module>moduleUi</module>
-        <module>moduleUtils</module>
-    </modules>
+    <packaging>jar</packaging>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -32,6 +31,9 @@
     </properties>
 
     <dependencies>
+
+
+
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-nop</artifactId>
@@ -129,9 +131,9 @@
                 </configuration>
                 <dependencies>
                     <dependency>
-                         <groupId>org.aspectj</groupId>
-                         <artifactId>aspectjweaver</artifactId>
-                         <version>${aspectj.version}</version>
+                        <groupId>org.aspectj</groupId>
+                        <artifactId>aspectjweaver</artifactId>
+                        <version>${aspectj.version}</version>
                     </dependency>
                 </dependencies>
             </plugin>
@@ -147,4 +149,5 @@
             </plugin>
         </plugins>
     </build>
+
 </project>
