+++ moduleUtils/src/main/resources/myConf.properties
@@ -1,0 +1,11 @@
+LOGIN =eXVsaWF0ZXN0aXJvdmE
+PASS =QXNkTGtqVmJuLTAxNQ
+DBHOST =localhost
+DBUSER =cm9vdA
+DBPASS =cm9vdA
+DBPORT =3306
+DBNAME =bGV0dGVyc3lhbmRleA
+CONTROLANSWER =MzQ0MDkz
+#LOGIN =eXVsaWEydGVzdDI
+#LOGIN =eXVsaWF0ZXN0aXJvdmE
+#LOGIN =eXVsaWExdGVzdA
