--- src/main/resources/myConf.properties
@@ -1,10 +1,0 @@
-LOGIN =eXVsaWF0ZXN0aXJvdmE
-PASS =QXNkTGtqVmJuLTAxNQ
-DBHOST =localhost
-DBUSER =cm9vdA
-DBPASS =cm9vdA
-DBPORT =3306
-DBNAME =bGV0dGVyc3lhbmRleA
-#LOGIN =eXVsaWEydGVzdDI
-#LOGIN =eXVsaWF0ZXN0aXJvdmE
-#LOGIN =eXVsaWExdGVzdA
