--- src/test/java/utils/apiUtils/modelMessage/Data.java
@@ -1,25 +1,0 @@
-package utils.apiUtils.modelMessage;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import java.util.ArrayList;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Data {
-    public ArrayList<Message> message;
-
-    @Override
-    public String toString() {
-        return "Data{" +
-                "message=" + message +
-                '}';
-    }
-
-    public ArrayList<Message> getMessage() {
-        return message;
-    }
-
-    public void setMessage(ArrayList<Message> message) {
-        this.message = message;
-    }
-}
